Added splashscreen, rewrite of how display images are generated, added ADC reading task

+++ firmware/Parts.cylib/ST7528/API/ST7528.h
@@ -12,7 +12,7 @@
 #include <CyLib.h>
 
 void `$INSTANCE_NAME`_Start();
-void `$INSTANCE_NAME`_WritePixels(uint8 data[], uint8 len);
+void `$INSTANCE_NAME`_WritePixels(uint8 data[], int len);
 void `$INSTANCE_NAME`_SetCursorPosition(uint8 page, uint8 col);
 void `$INSTANCE_NAME`_SetContrast(uint8 contrast_level);
 void `$INSTANCE_NAME`_DrawText(uint8 start_page, uint8 start_col, char *text, uint8 inverse);
